# fix(tsi1): optimize the comparison of SeriesIDSet.