--- a/tsdb/index/tsi1/log_file.go
+++ b/tsdb/index/tsi1/log_file.go
@@ -285,13 +285,7 @@ func (f *LogFile) MeasurementHasSeries(ss *tsdb.SeriesIDSet, name []byte) bool {
 		return false
 	}
 
-	// TODO(edd): if mm is using a seriesSet then this could be changed to do a fast intersection.
-	for _, id := range mm.seriesIDs() {
-		if ss.Contains(id) {
-			return true
-		}
-	}
-	return false
+	return mm.hasSeries(ss)
 }
 
 // MeasurementNames returns an ordered list of measurement names.
@@ -680,7 +674,7 @@ func (f *LogFile) execSeriesEntry(e *LogEntry) {
 	// the entire database and the server is restarted. This would cause
 	// the log to replay its insert but the key cannot be found.
 	//
-	// https://github.com/influxdata/influxdb/v2/issues/9444
+	// https://github.com/influxdata/influxdb/issues/9444
 	if seriesKey == nil {
 		return
 	}
@@ -1298,6 +1292,20 @@ func (m *logMeasurement) seriesIDSet() *tsdb.SeriesIDSet {
 		ss.AddNoLock(seriesID)
 	}
 	return ss
+}
+
+func (m *logMeasurement) hasSeries(ss *tsdb.SeriesIDSet) bool {
+	if m.seriesSet != nil {
+		return m.seriesSet.Intersects(ss)
+	}
+
+	for seriesID := range m.series {
+		if ss.Contains(seriesID) {
+			return true
+		}
+	}
+
+	return false
 }
 
 func (m *logMeasurement) Name() []byte  { return m.name }