(*LogFile).MeasurementHasSeries left a TODO, which is optimized with the intersection operation of bitmap.

The benchmark results before and after optimization are compared as follows:

|             | Before ns/op | Optimized ns/op | Before B/op | Optimized B/op | Before allocs/op | Optimized allocs/op |
| ----------- | ------------ | ------------ | ---------- | ----------- | ---------------- | ---------------- |
| 100 series  | 742          | 58           | 496        | 0           | 3                | 0                |
| 1k series   | 10862        | 57           | 8272       | 0           | 3                | 0                |
| 10k series  | 139394       | 61           | 8985       | 0           | 3                | 0                |
| 100k series | 1465113      | 57           | 802915     | 0           | 4                | 0                |

Theoretically, the time complexity is optimized from O(n) to O(1), and the space complexity is optimized from O(n) to O(1). The test results are basically responsible for theoretical prediction.

<!-- Checkboxes below this note can be erased if not applicable to your Pull Request. -->

- [ ] [CHANGELOG.md](https://github.com/influxdata/influxdb/blob/master/CHANGELOG.md) updated with a link to the PR (not the Issue)
- [ ] [Well-formatted commit messages](https://www.conventionalcommits.org/en/v1.0.0-beta.3/)
- [ ] Rebased/mergeable
- [ ] Tests pass
- [ ] Signed [CLA](https://influxdata.com/community/cla/) (if not already signed)